+++ src/card/sets/y2011/innistrad.rs
@@ -348,6 +348,7 @@ pub(in crate::card::sets) static FIEND_HUNTER: CardRecord = CardRecord::new(
                 EffectDef::ReturnLinkedExiles {
                     zone: ZoneKind::Battlefield,
                     grant: None,
+                    controller: None,
                 },
             ),
         ]),