+++ src/card/sets/y2013/dragons_maze.rs
@@ -274,6 +274,7 @@ pub(in crate::card::sets) static AETHERLING: CardRecord = CardRecord::new(
                     EffectDef::ReturnLinkedExiles {
                         zone: ZoneKind::Battlefield,
                         grant: None,
+                        controller: None,
                     },
                 ))),
             ]),